+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - `manifold teams members` can list members of a team
+- `manifold teams invite` now supports Role selection
 
 ### Fixed